Starting authentication
This first implementation of the Hawk authentication scheme has a lot of hard-coded data that I mean to get rid of in subsequent commits. Also, Eve currently doesn't have an internal method for `GET` requests, so there was no way I could bypass authentication for the HTML renderer.

The HTML renderer is currently behaving like any other external app would (note the `apps` folder that was added and the sample client). My real intention is that it actually be handled like a *renderer*. In order to do that it might be necessary to add some kind of HTML rendering class to Eve by looking at how the JSON and XML renderers work.

The `posts` endpoint is currently set to be completely private, but it should actually check if the requested resource has permissions set to public before going through the authentication.