Finalizing server authentication
The server will now create credentials posts automatically when app posts are created. Possible digits in a post ID had to be increased to 9 in order to support microsecond timestamps and Eve needed to be updated to the bleeding edge in order to support this.

This commit also introduces the concept of *root credentials* in order to be able to authenticate that first app, but it would be *extremely* unwise for apps to continue to use root credentials indefinitely. In any case, root credentials should only be used for apps that reside on the same machine as the PISS instance and should *never* be transmitted over the wire in any way, shape, or form. The credentials are saved in `piss.cfg` under `ROOT_CREDENTIALS`.

I've also decided that it's a waste of database lookups to have `meta` posts on the server, since PISS is designed to be single-entity only. As a result, `META_POST` was also added to `piss.cfg`.

Fixes #15.