A method for app self-registration
If no configuration file is found, the app will create an app post on the server, retrieve the app's credentials, and write those credentials to a config file. This prevents giving users an extra setup step where they need to supply root credentials first and then change the file to reflect the app's own credentials.

To achieve this, I changed the server to expose the post type for every successful `POST`. In addition, an event hook is fired after every `POST` to consume the post type and determine whether it should add a credentials link to the header.

Fixes #16.