Switching `types` endpoint from Eve to files
The `types` endpoint will be used to load the contents of a specific schema
to memory in order to perform validations. It isn't necessary to search the
contents of a type -- they will always be accessed by name. In addition, type
files are generally quite small, so the entire contents can be loaded into
memory at once. Types should be added very deliberately and will very rarely
need to be modified once they have been created. As a result, the `types`
endpoint is better off reading from the file system than accessing a database.

Keeping the `types` endpoint on the file system will also allow me to add the
files to access control to keep track of changes to the schema.

This commit implements all the changes necessary to switch the `types` endpoint
from a database endpoint to a file-system endpoint.