Started work on themes
The general idea is that there will be an internal 'default' theme that can be
added to / overridden by user-generated themes. Specifying the name of the
theme in `piss.cfg` will add that folder to the internal template lookup. If a
particular file isn't found in the custom theme, the server will automatically
use a default template in its place. This will allow users to only create files
for the specific parts of the template they'd like to change (though they can
override the whole thing if desired).