Syndication service
Fleshed out how syndication services will be handled and added a
Twitter service.

Apps POST an object to the syndication endpoint of the desired service.
The object has an identical structure to a PISS post, however:

  * `entity` is set to the service URL
  * `type` is set to the post type the service can handle
  * `content` key contains the data to be syndicated
  * `links` has a single element with the `id` of the original post

The server will then check if a syndicated link to the same service
exists (duplicates will be rejected), create the syndicated post on
the selected service's API, then PATCH the original post to include
the syndicated link in the `links` list.

Having the server perform the actual syndication to various services
allows the service credentials to reside in a single place and it
would allow users to later specify which apps are allowed to syndicate.

Fixes #29.