Templates now subdivided based on post type
A new `types` directory has been added to the templates so that
individual styling can be added on a per-post-type basis. This
should still fall back to sensible defaults if the current theme
doesn't have a style for a specific post type.

To accomplish this change, I've separated out the Jinja2 code from
`eve_override.py` into its own `jinja_override.py` file. I've also
added a few template API functions in order to retrieve the title
and body of a post for single pages and feeds.